Add PDF output format to deep research reports
The deep research pipeline could only produce Markdown and HTML reports.
This commit adds PDF as a third output format using picoloom
(github.com/alnah/picoloom/v2), a Markdown-to-PDF converter that uses
Chrome headless.

On the backend, the State struct now carries three report rendering slots
(MarkdownReport, HTMLReport, PDFReport) tagged json:"-" since all three
are produced by the final Reporter node and never read by downstream
graph nodes. The generateTraditionalReport and generateChapterBasedReport
functions only render the format that DeepResearchConfig.ReportFormat
requests instead of generating both HTML and PDF unconditionally.

The Attachment.Size field is widened from int to int64 to match its ES
mapping (type:long), and saveReport accepts both text and binary content
via mutually exclusive report/reportBytes parameters.

On the frontend, the assistant editor now includes PDF alongside Markdown
and HTML in the report format dropdown.

Also, service.GetAssistant error handling is split across three call sites
in session.go: a fetch error now returns 500 "failed to get assistant"
while a genuinely missing assistant still returns the not-found response.

+++ core/attachment.go
@@ -14,7 +14,7 @@ type Attachment struct {
 	Icon        string `json:"icon,omitempty" elastic_mapping:"icon:{enabled:false}"`
 	MimeType    string `json:"mime_type,omitempty" elastic_mapping:"mime_type:{enabled:false}"`
 	URL         string `json:"url,omitempty" elastic_mapping:"url:{enabled:false}"`
-	Size        int    `json:"size,omitempty" elastic_mapping:"size:{type:long}"`
+	Size        int64  `json:"size,omitempty" elastic_mapping:"size:{type:long}"`
 	Text        string `json:"text,omitempty" elastic_mapping:"text:{type:text}"` // Extracted text content (e.g., via OCR)
 
 	Deleted       bool        `json:"deleted,omitempty" elastic_mapping:"deleted:{type:boolean}"`
